+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -7,7 +7,8 @@
     ([#23486](https://github.com/Azure/azure-sdk-for-python/pull/23486))
 - Implement sending of exception telemetry via log exporter
     ([#23633](https://github.com/Azure/azure-sdk-for-python/pull/23633))
-
+- Implement exporting span events as message/exception telemetry
+    ([#23633](https://github.com/Azure/azure-sdk-for-python/pull/23633))
 
 ### Breaking Changes